+++ R/utils_path.R
@@ -34,6 +34,18 @@ path_objects <- function(path_store, name) {
   file.path(path_objects_dir(path_store), name)
 }
 
+#' @title Path to directory of saved targets
+#' @export
+#' @keywords internal
+#' @description Internal function. Not for users.
+#' @param path_store Path to the data store.
+#' @param name Target names.
+#' @examples
+#' tar_path_objects_dir("_targets")
+tar_path_objects_dir <- function(path_store) {
+  path_objects_dir(path_store = path_store)
+}
+
 path_objects_dir <- function(path_store) {
   file.path(path_store, "objects")
 }